Refactor frost into round1, round2 submodules
