+++ frost-ristretto255/src/frost/keys.rs
@@ -14,7 +14,7 @@ use crate::VerificationKey;
 
 /// A secret scalar value representing a single signer's secret key.
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
-pub struct Secret(pub(crate) Scalar);
+pub struct Secret(pub(super) Scalar);
 
 impl Secret {
     /// Generates a new uniformly random secret value using the provided RNG.
@@ -62,7 +62,7 @@ impl TryFrom<[u8; 32]> for Secret {
 
 /// A public group element that represents a single signer's public key.
 #[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Public(pub(crate) RistrettoPoint);
+pub struct Public(pub(super) RistrettoPoint);
 
 impl From<RistrettoPoint> for Public {
     fn from(source: RistrettoPoint) -> Public {
@@ -82,7 +82,7 @@ impl From<Secret> for Public {
 /// This is a (public) commitment to one coefficient of a secret polynomial used
 /// for performing verifiable secret sharing for a Shamir secret share.
 #[derive(Clone, Copy, Debug, PartialEq)]
-pub(crate) struct CoefficientCommitment(pub(crate) RistrettoPoint);
+pub(super) struct CoefficientCommitment(pub(super) RistrettoPoint);
 
 /// Contains the commitments to the coefficients for our secret polynomial _f_,
 /// used to generate participants' key shares.
@@ -97,7 +97,7 @@ pub(crate) struct CoefficientCommitment(pub(crate) RistrettoPoint);
 /// some agreed-upon public location for publication, where each participant can
 /// ensure that they received the correct (and same) value.
 #[derive(Clone)]
-pub struct VerifiableSecretSharingCommitment(pub(crate) Vec<CoefficientCommitment>);
+pub struct VerifiableSecretSharingCommitment(pub(super) Vec<CoefficientCommitment>);
 
 /// A secret share generated by performing a (t-out-of-n) secret sharing scheme.
 ///
@@ -105,11 +105,11 @@ pub struct VerifiableSecretSharingCommitment(pub(crate) Vec<CoefficientCommitmen
 /// in this case we use Shamir's secret sharing.
 #[derive(Clone)]
 pub struct SecretShare {
-    pub(crate) index: u16,
+    pub(super) index: u16,
     /// Secret Key.
-    pub(crate) value: Secret,
+    pub(super) value: Secret,
     /// The commitments to be distributed among signers.
-    pub(crate) commitment: VerifiableSecretSharingCommitment,
+    pub(super) commitment: VerifiableSecretSharingCommitment,
 }
 
 impl SecretShare {
@@ -144,7 +144,7 @@ impl SecretShare {
 /// for performing verifiable secret sharing for a Shamir secret share.
 #[derive(Clone, Copy, Debug, PartialEq)]
 // TODO: deprecate
-pub(crate) struct Commitment(pub(crate) RistrettoPoint);
+pub(super) struct Commitment(pub(super) RistrettoPoint);
 
 /// Secret and public key material generated by a dealer performing
 /// [`keygen_with_dealer`].
@@ -156,7 +156,7 @@ pub struct SharePackage {
     /// Denotes the participant index each share is owned by.
     pub index: u16,
     /// This participant's secret share.
-    pub(crate) secret_share: SecretShare,
+    pub(super) secret_share: SecretShare,
     /// This participant's public key.
     pub public: Public,
     /// The public signing key that represents the entire group.
@@ -218,7 +218,7 @@ pub struct KeyPackage {
     /// Denotes the participant index each secret share key package is owned by.
     pub index: u16,
     /// This participant's secret share.
-    pub(crate) secret_share: Secret,
+    pub(super) secret_share: Secret,
     /// This participant's public key.
     pub public: Public,
     /// The public signing key that represents the entire group.
@@ -257,7 +257,7 @@ pub struct PublicKeyPackage {
     /// correct view of participants' public keys to perform verification before
     /// publishing a signature. `signer_pubkeys` represents all signers for a
     /// signing operation.
-    pub(crate) signer_pubkeys: HashMap<u16, Public>,
+    pub(super) signer_pubkeys: HashMap<u16, Public>,
     /// The joint public key for the entire group.
     pub group_public: VerificationKey,
 }
