+++ frost-ristretto255/tests/frost.rs
@@ -18,8 +18,8 @@ fn check_sign_with_dealer() {
         .map(|share| frost::keys::KeyPackage::try_from(share).unwrap())
         .collect();
 
-    let mut nonces: HashMap<u16, Vec<frost::SigningNonces>> = HashMap::new();
-    let mut commitments: HashMap<u16, Vec<frost::SigningCommitments>> = HashMap::new();
+    let mut nonces: HashMap<u16, Vec<frost::round1::SigningNonces>> = HashMap::new();
+    let mut commitments: HashMap<u16, Vec<frost::round1::SigningCommitments>> = HashMap::new();
 
     ////////////////////////////////////////////////////////////////////////////
     // Round 1: generating nonces and signing commitments for each participant
@@ -28,7 +28,7 @@ fn check_sign_with_dealer() {
     for participant_index in 1..(threshold + 1) {
         // Generate one (1) nonce and one SigningCommitments instance for each
         // participant, up to _threshold_.
-        let (nonce, commitment) = frost::preprocess(1, participant_index as u16, &mut rng);
+        let (nonce, commitment) = frost::round1::preprocess(1, participant_index as u16, &mut rng);
         nonces.insert(participant_index as u16, nonce);
         commitments.insert(participant_index as u16, commitment);
     }
@@ -36,10 +36,10 @@ fn check_sign_with_dealer() {
     // This is what the signature aggregator / coordinator needs to do:
     // - decide what message to sign
     // - take one (unused) commitment per signing participant
-    let mut signature_shares: Vec<frost::SignatureShare> = Vec::new();
+    let mut signature_shares: Vec<frost::round2::SignatureShare> = Vec::new();
     let message = "message to sign".as_bytes();
     let comms = commitments.clone().into_values().flatten().collect();
-    let signing_package = frost::SigningPackage::new(comms, message.to_vec());
+    let signing_package = frost::round2::SigningPackage::new(comms, message.to_vec());
 
     ////////////////////////////////////////////////////////////////////////////
     // Round 2: each participant generates their signature share
@@ -55,7 +55,7 @@ fn check_sign_with_dealer() {
         let commitments_to_use = commitments.get(participant_index).unwrap()[0];
 
         // Each participant generates their signature share.
-        let signature_share = frost::sign(
+        let signature_share = frost::round2::sign(
             &signing_package,
             &nonces_to_use,
             &commitments_to_use,
